.travis-requirements.txt
behave
coverage
coveralls
nose
PyHamcrest
scrutinizer-ocular
six
